refactor: 가독성 개선을 위한 메서드 분리
- 불필요한 쿼리를 제거합니다.
