+++ src/main/java/roomescape/application/reservation/query/WaitingQueryService.java
@@ -5,32 +5,28 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.application.reservation.query.dto.WaitingResult;
 import roomescape.application.reservation.query.dto.WaitingWithRankResult;
-import roomescape.domain.member.Member;
-import roomescape.domain.member.repository.MemberRepository;
 import roomescape.domain.reservation.WaitingWithRank;
 import roomescape.domain.reservation.repository.WaitingRepository;
-import roomescape.infrastructure.error.exception.ReservationException;
 
 @Service
 @Transactional(readOnly = true)
 public class WaitingQueryService {
 
     private final WaitingRepository waitingRepository;
-    private final MemberRepository memberRepository;
 
-    public WaitingQueryService(WaitingRepository waitingRepository, MemberRepository memberRepository) {
+    public WaitingQueryService(WaitingRepository waitingRepository) {
         this.waitingRepository = waitingRepository;
-        this.memberRepository = memberRepository;
     }
 
     public List<WaitingWithRankResult> findWaitingByMemberId(Long memberId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new ReservationException("회원 정보가 존재하지 않습니다."));
-        List<WaitingWithRank> waitings = waitingRepository.findWaitingsWithRankByMemberId(member.getId());
+        List<WaitingWithRank> waitings = waitingRepository.findWaitingsWithRankByMemberId(memberId);
         return waitings.stream()
-                .map(waitingWithRank -> new WaitingWithRank(waitingWithRank.waiting(), waitingWithRank.rank() + 1))
-                .map(waitingWithRank -> WaitingWithRankResult.from(waitingWithRank.waiting(), waitingWithRank.rank()))
+                .map(w -> WaitingWithRankResult.from(w.waiting(), oneBasedRank(w.rank())))
                 .toList();
+    }
+
+    private long oneBasedRank(long zeroBasedRank) {
+        return zeroBasedRank + 1;
     }
 
     public List<WaitingResult> findAll() {
